refactor: clarify credit notification type editing
Changed:
Reworked the credit notification type editor into basic settings, SMS template, and trigger rule sections with a clear done action and draft-save hint.

Benefit:
Operators can complete notification type edits with clearer structure while the existing save flow and policy payload stay unchanged.
