+++ src/i18n/ar-SA/modules/operations-credit-notifications.json
@@ -215,6 +215,24 @@
       "credit_granted": "تنبيه المعلم للتحقق من وصول الأرصدة بعد إضافتها.",
       "low_balance": "تنبيه المعلم عند انخفاض رصيد الأرصدة عن الحد أو عدم كفاية الأيام المتوقعة المتاحة."
     },
+    "typeEditor": {
+      "descriptions": {
+        "basic": "تحكّم فيما إذا كان هذا النوع من الإشعارات سيشارك في عمليات الإرسال القادمة.",
+        "template": "اختر قالب الرسائل النصية أو تحقق منه. يبقى دليل المتغيرات مطويًا إلى أن تحتاج إلى فحص القالب."
+      },
+      "done": "إنهاء التحرير",
+      "draftHint": "تبقى التغييرات أولًا في مسودة هذه الصفحة. اضغط زر حفظ الإعدادات في الأسفل لتفعيلها.",
+      "ruleDescriptions": {
+        "credit_expiring": "اضبط مواعيد تذكير المعلمين قبل انتهاء الرصيد، وما إذا كان سيتم دمج عدة أرصدة منتهية لكل معلم.",
+        "credit_granted": "يتم تشغيل هذا الإشعار بعد إضافة الرصيد ولا يحتاج إلى إعدادات تشغيل إضافية.",
+        "low_balance": "اضبط حدود الرصيد الثابتة أو ذكّر المعلمين حسب عدد أيام الاستخدام المتبقية المقدّرة."
+      },
+      "sections": {
+        "basic": "الإعدادات الأساسية",
+        "rules": "قواعد التشغيل",
+        "template": "قالب الرسائل النصية"
+      }
+    },
     "typeTable": {
       "actions": "الإجراءات",
       "channel": "القناة",
