+++ src/i18n/th-TH/modules/operations-credit-notifications.json
@@ -215,6 +215,24 @@
       "credit_granted": "เตือนครูให้ตรวจสอบเครดิตเข้าหลังได้รับเครดิต",
       "low_balance": "เตือนครูเมื่อเครดิตต่ำกว่าเกณฑ์หรือจำนวนวันที่คาดว่าจะใช้งานได้ไม่พอ"
     },
+    "typeEditor": {
+      "descriptions": {
+        "basic": "กำหนดว่าการแจ้งเตือนประเภทนี้จะใช้กับการส่งครั้งต่อไปหรือไม่",
+        "template": "เลือกหรือตรวจสอบเทมเพลต SMS โดยคำแนะนำตัวแปรจะถูกพับไว้จนกว่าจะต้องตรวจสอบเทมเพลต"
+      },
+      "done": "แก้ไขเสร็จแล้ว",
+      "draftHint": "การเปลี่ยนแปลงจะถูกเก็บไว้ในแบบร่างของหน้านี้ก่อน คลิกปุ่มบันทึกการตั้งค่าด้านล่างเพื่อให้มีผล",
+      "ruleDescriptions": {
+        "credit_expiring": "กำหนดเวลาที่จะเตือนครูก่อนเครดิตหมดอายุ และกำหนดว่าจะรวมเครดิตที่ใกล้หมดอายุหลายรายการของครูคนเดียวกันหรือไม่",
+        "credit_granted": "การแจ้งเตือนนี้จะทำงานหลังจากเติมเครดิตสำเร็จ และไม่ต้องตั้งค่าเงื่อนไขเพิ่มเติม",
+        "low_balance": "กำหนดเกณฑ์ยอดเครดิตคงเหลือแบบตายตัว หรือเตือนครูตามจำนวนวันใช้งานที่เหลือโดยประมาณ"
+      },
+      "sections": {
+        "basic": "การตั้งค่าพื้นฐาน",
+        "rules": "กฎการแจ้งเตือน",
+        "template": "เทมเพลต SMS"
+      }
+    },
     "typeTable": {
       "actions": "การดำเนินการ",
       "channel": "ช่องทาง",
